Add support for reading updates that are split across multiple files
These contain a SquashFS file system that itself contains the .spk file.
I'm not clear why they use that additional layer rather than splitting
the .spk file directly.

This implementation uses the backhand crate for working with squashfs.
It contains two inefficiencies: the files must be loaded into memory so
that backhand can process it (it wants something that implements
BufReadSeek), and the file data must be read from SquashFS into a new
buffer before it can be processed because the SquashfsReadFile does not
implement Seek.

The former can likely be resolved by implementing a type that wraps
multiple file objects and implements Read + Seek in terms of them. The
latter would require changes to backhand.